… examples/filter-simple/main.wasm
… examples/score-simple/main.wasm
… guest/testdata/noop/main.wasm