#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

Updates to latest TinyGo which progresses its applicability to things including json and ASN code.

#### Which issue(s) this PR fixes:

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?

Minimum TinyGo is now latest TinyGo 0.28.1 (technically 0.28.0 is fine)

#### What are the benchmark results of this change?

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
                                       │ v0.27.0.txt  │            v0.28.1.txt             │
                                       │    sec/op    │   sec/op     vs base               │
PluginFilter/noop/params:_small-12       199.7n ± 14%   189.6n ± 2%   -5.06% (p=0.002 n=6)
PluginFilter/noop/params:_real-12        191.7n ±  3%   189.2n ± 1%   -1.30% (p=0.002 n=6)
PluginFilter/wasm/params:_small-12       6.869µ ±  2%   6.844µ ± 1%        ~ (p=0.180 n=6)
PluginFilter/wasm/params:_real-12        50.39µ ±  1%   50.57µ ± 1%        ~ (p=0.394 n=6)
PluginScore/noop/params:_small-12        241.8n ±  1%   243.2n ± 1%        ~ (p=0.327 n=6)
PluginScore/noop/params:_real-12         206.4n ±  1%   207.9n ± 0%   +0.73% (p=0.002 n=6)
PluginScore/wasm/params:_small-12        5.158µ ±  0%   3.918µ ± 1%  -24.03% (p=0.002 n=6)
PluginScore/wasm/params:_real-12         49.36µ ±  2%   51.76µ ± 2%   +4.85% (p=0.002 n=6)

```