@codefromthecrypt 2023-06-12
> Can we somehow tell users about the recommended tinygo version?

yeah. actually we are missing any info at all about building the guest, so when we make that we should also tell them about the tinygo version ;) I've personally delayed this until the functionality is a little farther because compiling guest locks in the ABI, which is moving still.